+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - support for pandas versions 1.5 - 2.0
 - HOTFIX: mkdir called on destination instead of default location in AlyxClient.download_cache_tables
 - date field of session details a datetime.date object in remote mode (now consistent with local mode)
+- support datasets table without session_path field
 
 ## [1.21.3]